moved created by ref to global, made it compatible with assessments 2.0, added it to assessments 2.0
