+++ src/app/assess/wizard/wizard.component.html
@@ -141,6 +141,11 @@
             </mat-form-field>
           </div>
         </div>
+        <div class="row mat-card-padding">
+          <div class="col-xs-12">
+            <created-by-ref [assessmentMeta]="meta"></created-by-ref>
+          </div>
+        </div>
       </mat-card-content>
     </ng-template>
