+++ src/app/global/form-models/assess.ts
@@ -10,6 +10,7 @@ export const AssessForm = (assessment?: AssessmentMeta) => {
     return new FormGroup({
         title: new FormControl(assessment.title || '', Validators.required),
         description: new FormControl(assessment.description || ''),
+        created_by_ref: new FormControl(assessment.created_by_ref),
         includesIndicators: new FormControl(assessment.includesIndicators),
         includesMitigations: new FormControl(assessment.includesMitigations),
         includesSensors: new FormControl(assessment.includesSensors),
