+++ src/app/global/global.module.ts
@@ -14,7 +14,7 @@ import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { MatTableModule } from '@angular/material/table';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { MatToolbarModule } from '@angular/material/toolbar';
-import { MatListModule, MatDialog, MatCardModule } from '@angular/material';
+import { MatListModule, MatDialog, MatCardModule, MatSelectModule } from '@angular/material';
 import { MatMenuModule } from '@angular/material/menu';
 import { MatExpansionModule } from '@angular/material';
 import { MatSidenavModule } from '@angular/material/sidenav';
@@ -57,6 +57,7 @@ import { ErrorCardComponent } from './components/error-card/error-card.component
 import { RiskBreakdownComponent } from './components/risk-breakdown/risk-breakdown.component';
 import { TreemapComponent } from './components/treemap/treemap.component';
 import { HeatmapComponent } from './components/heatmap/heatmap.component';
+import { CreatedByRefComponent } from './components/created-by-ref/created-by-ref.component';
 
 const matModules = [
     MatAutocompleteModule,
@@ -76,7 +77,8 @@ const matModules = [
     MatExpansionModule,
     MatDialogModule,
     MatTabsModule,
-    MatCardModule
+    MatCardModule,
+    MatSelectModule
 ];
 
 const unfetterComponents = [
@@ -110,6 +112,7 @@ const unfetterComponents = [
     RiskBreakdownComponent,
     TreemapComponent,
     HeatmapComponent,
+    CreatedByRefComponent,
 ];
 
 @NgModule({
