+++ src/app/models/assess/assessment-meta.ts
@@ -2,6 +2,7 @@
 export class AssessmentMeta {
     public title = '';
     public description = '';
+    public created_by_ref = '';
     public includesIndicators = false;
     public includesMitigations = false;
     public includesSensors = false;
